requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
openai-whisper
torch
torchaudio
ffmpeg-python==0.2.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy<2
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0